version 1.1

+++ BasicFileEncoder/src/headers/FileManager.h
@@ -1,4 +1,4 @@
 #pragma once
-int fileExtract(const char* location, unsigned char*& fileBuffer, unsigned long& bufferLength);
+int fileExtract(const char* location, unsigned char*& fileBuffer, unsigned long long& bufferLength);
 
-int fileWritter(const char* name, unsigned char* fileBuffer, unsigned long bufferLength);
+int fileWritter(const char* name, unsigned char* fileBuffer, unsigned long long bufferLength);
